+++ pandapower/pf/create_jacobian.py
@@ -2,7 +2,7 @@ from numpy import complex128, float64, int32
 from numpy.core.multiarray import zeros, empty, array
 from scipy.sparse import csr_matrix as sparse, vstack, hstack
 
-from pandapower.pf.dSbus_dV_pypower import dSbus_dV
+from pandapower.pypower.dSbus_dV import dSbus_dV
 
 try:
     # numba functions