+++ pandapower/results_gen.py
@@ -1,12 +1,12 @@
 # -*- coding: utf-8 -*-
 
-# Copyright (c) 2016-2018 by University of Kassel and Fraunhofer Institute for Energy Economics
+# Copyright (c) 2016-2019 by University of Kassel and Fraunhofer Institute for Energy Economics
 # and Energy System Technology (IEE), Kassel. All rights reserved.
 
 
 import numpy as np
-from pandapower.idx_bus import VM, VA
-from pandapower.idx_gen import PG, QG
+from pandapower.pypower.idx_bus import VM, VA
+from pandapower.pypower.idx_gen import PG, QG
 
 from pandapower.auxiliary import _sum_by_group